feat: 도서 상태 api연동, 리랜더링 이슈로 책 정보에 관한 컴포넌트 세부화 작업
